openai-agent-practice/hello-agent/pyproject.toml
[project]
name = "hello-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "wajidminhas", email = "shanitent667@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "chainlit>=2.5.5",
    "google-generativeai>=0.8.5",
    "openai>=1.86.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]

[project.scripts]
hello-agent = "hello_agent:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
